questionGenerationApi/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
pydealer==1.4.0